+++ test/main.js
@@ -1,5 +1,0 @@
-import './fixtures/v0.4.7/input'
-
-document.querySelector('#app').innerHTML = `
-  <h1>Hey Vite 👋</h1>
-`