+++ test/src/exports.js
@@ -1,0 +1,11 @@
+
+const { hello, world } = require('./dynamic')
+
+exports.msg = `
+[foo.js]
+
+const { hello, world } = require('./dynamic')
+
+hello: ${hello}
+world: ${world}
+`