+++ test/src/main.ts
@@ -1,0 +1,7 @@
+const { msg: message } = require('./exports')
+
+document.querySelector('#app')!.innerHTML = `
+  <pre>
+    ${message}
+  </pre>
+`