+++ test/src/module-exports/world.cjs
@@ -1,0 +1,2 @@
+
+module.exports = 'module-exports/world.cjs'